requirements.txt
Flask==3.1.2
PyPDF2==3.0.1
python-docx==1.2.0
scikit-learn==1.7.2
spacy==3.8.9
openai==1.54.0
transformers==4.46.0
reportlab==4.2.5
matplotlib==3.9.2
flask-login==0.7.0
flask-wtf==1.2.2
flask-sqlalchemy==3.1.1
werkzeug==3.1.3
flask-bcrypt==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
langdetect==1.0.9
flask-babel==4.0.0
pillow==11.0.0
python-dotenv==1.0.1
thinc==8.3.2
numpy==1.26.4
